# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "D&D 5E Character Creator API",
        "description": "RESTful API for creating and managing D&D 5th Edition (2024 ruleset) characters. Provides endpoints for character CRUD operations and reference data for classes, species, and backgrounds.",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/classes": {
            "get": {
                "tags": [
                    "Classes"
                ],
                "summary": "List all classes",
                "description": "Returns a paginated list of D&D character classes.",
                "operationId": "fdaa3ee819d965230e7e81af690b7efd",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Items per page (max 100)",
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "filter.spellcaster",
                        "in": "query",
                        "description": "Filter to only spellcasting classes",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "true"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Classes retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Class"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Classes retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/classes/{id}": {
            "get": {
                "tags": [
                    "Classes"
                ],
                "summary": "Get a class",
                "description": "Returns detailed information about a D&D class including subclasses and features.",
                "operationId": "ce167415b4a87811fb6e47c908e55c84",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Class UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Include features",
                        "schema": {
                            "type": "string",
                            "example": "features"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Class retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Class"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Class retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Class not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/species": {
            "get": {
                "tags": [
                    "Species"
                ],
                "summary": "List all species",
                "description": "Returns a paginated list of D&D character species.",
                "operationId": "95f7879f7a35e76efee25b7121fcbeb4",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Items per page (max 100)",
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Species retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Species"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Species retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/species/{id}": {
            "get": {
                "tags": [
                    "Species"
                ],
                "summary": "Get a species",
                "description": "Returns detailed information about a D&D species including traits.",
                "operationId": "9cdcacd76fb3a8baaf39522430aeaef2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Species UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Species retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Species"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Species retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Species not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Class": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/PaginationMeta"
                    },
                    {
                        "properties": {
                            "id": {
                                "description": "Read-only controller for class reference data.",
                                "type": "string",
                                "format": "uuid"
                            },
                            "name": {
                                "type": "string",
                                "example": "Fighter"
                            },
                            "slug": {
                                "type": "string",
                                "example": "fighter"
                            },
                            "description": {
                                "type": "string"
                            },
                            "hitDie": {
                                "type": "integer",
                                "example": 10
                            },
                            "primaryAbilities": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "savingThrowProficiencies": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "armorProficiencies": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "weaponProficiencies": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "toolProficiencies": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "nullable": true
                            },
                            "skillChoices": {
                                "type": "object"
                            },
                            "startingEquipment": {
                                "type": "object"
                            },
                            "spellcastingAbility": {
                                "type": "string",
                                "nullable": true
                            },
                            "subclassLevel": {
                                "type": "integer",
                                "example": 3
                            },
                            "subclasses": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Subclass"
                                }
                            },
                            "features": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/ClassFeature"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "Subclass": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/PaginationMeta"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "name": {
                                "type": "string",
                                "example": "Champion"
                            },
                            "slug": {
                                "type": "string",
                                "example": "champion"
                            },
                            "description": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "ClassFeature": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/PaginationMeta"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "name": {
                                "type": "string",
                                "example": "Second Wind"
                            },
                            "level": {
                                "type": "integer",
                                "example": 1
                            },
                            "description": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "ErrorResponse": {
                "required": [
                    "success",
                    "error",
                    "meta"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "required": [
                            "code",
                            "message"
                        ],
                        "properties": {
                            "code": {
                                "type": "string",
                                "example": "VALIDATION_ERROR"
                            },
                            "message": {
                                "type": "string",
                                "example": "The given data was invalid."
                            },
                            "details": {
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "timestamp": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "correlationId": {
                                "type": "string",
                                "format": "uuid"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "currentPage": {
                        "type": "integer",
                        "example": 1
                    },
                    "perPage": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 42
                    },
                    "lastPage": {
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            },
            "Species": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/PaginationMeta"
                    },
                    {
                        "properties": {
                            "id": {
                                "description": "Read-only controller for species reference data.",
                                "type": "string",
                                "format": "uuid"
                            },
                            "name": {
                                "type": "string",
                                "example": "Human"
                            },
                            "slug": {
                                "type": "string",
                                "example": "human"
                            },
                            "description": {
                                "type": "string"
                            },
                            "size": {
                                "type": "string",
                                "example": "Medium"
                            },
                            "speed": {
                                "type": "integer",
                                "example": 30
                            },
                            "creatureType": {
                                "type": "string",
                                "example": "Humanoid"
                            },
                            "darkvision": {
                                "type": "integer",
                                "example": 60,
                                "nullable": true
                            },
                            "traits": {
                                "type": "array",
                                "items": {
                                    "type": "object"
                                }
                            },
                            "languages": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "abilityScoreOptions": {
                                "type": "object"
                            },
                            "hasLineageChoice": {
                                "type": "boolean"
                            },
                            "lineages": {
                                "type": "array",
                                "items": {
                                    "type": "object"
                                },
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                ]
            }
        }
    },
...
}